[Add] Support for Standardize Layout
