+++ main.go
@@ -1,14 +1,11 @@
 package main
 
 import (
-    "github.com/hashicorp/terraform/terraform"
-    "github.com/hashicorp/terraform/plugin"
+	"github.com/Cardfree/terraform-provider-teamcity/teamcity"
+	"github.com/hashicorp/terraform/plugin"
 )
 
 func main() {
-    plugin.Serve(&plugin.ServeOpts{
-        ProviderFunc: func() terraform.ResourceProvider {
-            return Provider()
-        },
-    })
+	plugin.Serve(&plugin.ServeOpts{
+		ProviderFunc: teamcity.Provider})
 }
