+++ teamcity/resource_build_configuration.go
@@ -1,4 +1,4 @@
-package main
+package teamcity
 
 import (
 	// "errors"
