+++ teamcity/resource_vcs_root.go
@@ -1,4 +1,4 @@
-package main
+package teamcity
 
 import (
 	// "fmt"
