+++ teamcity/resource_vcs_root_test.go
@@ -1,4 +1,4 @@
-package main
+package teamcity
 
 import (
 	"fmt"
